requirements.txt
rs3_api
python-osrsapi
customtkinter
aiohttp